+++ R/bioconductor.R
@@ -8,7 +8,7 @@ renv_bioconductor_init <- function(library = NULL) {
 
 renv_bioconductor_init_biocmanager <- function(library = NULL) {
 
-  if ("BiocManager" %in% loadedNamespaces())
+  if (renv_package_checking())
     return(TRUE)
 
   library <- library %||% renv_libpaths_default()
@@ -23,7 +23,7 @@ renv_bioconductor_init_biocmanager <- function(library = NULL) {
 
 renv_bioconductor_init_biocinstaller <- function(library = NULL) {
 
-  if ("BiocInstaller" %in% loadedNamespaces())
+  if (renv_package_checking())
     return(TRUE)
 
   library <- library %||% renv_libpaths_default()