+++ R/envvars.R
@@ -56,14 +56,14 @@ renv_envvars_init <- function() {
 
   renv_envvars_normalize()
 
-  if (renv_platform_macos()) {
-
-    # set SDKROOT so older R installations can find command line tools
+  # set SDKROOT so older R installations can find command line tools
+  # this was necessary for R installations built using a custom LLVM toolchain
+  # which did not set '-isysroot <...>' within their etc/Makevars file
+  if (renv_platform_macos() && getRversion() < "4.0.0") {
     sdk <- "/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk"
     sdkroot <- Sys.getenv("SDKROOT", unset = NA)
     if (is.na(sdkroot) && file.exists(sdk))
       Sys.setenv(SDKROOT = sdk)
-
   }
 
 }